[ATR-209] feat: User와 UserDetail 분리
- 로그인 과정에서 받는 유저의 추가정보(UserDetail) User에서 분리
- UserDetail이라는 Entity
