+++ src/main/java/run/attraction/api/v1/user/repository/UserRepository.java
@@ -4,5 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import run.attraction.api.v1.user.User;
 
 public interface UserRepository extends JpaRepository<User, String> {
-  boolean existsByNickname(String nickname);
 }
